add time and button for updating deleting post

+++ src/app/components/posts/posts.component.ts
@@ -1,3 +1,4 @@
+// posts.component.ts
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
@@ -6,20 +7,23 @@ import { PostActions } from '../../store/posts/post.actions';
 import { AuthService } from '../../services/auth.service';
 import { selectFilteredPosts } from '../../store/posts/post.reducer';
 import { FormsModule } from '@angular/forms';
-import { CommonModule } from '@angular/common'; 
+import { CommonModule } from '@angular/common';
 import { NavbarComponent } from '../../layout/navbar/navbar.component';
+import Swal from 'sweetalert2';
+import { format } from 'date-fns'; // Import the format function
 
 @Component({
   selector: 'app-posts',
   standalone: true,
-  imports: [ FormsModule, CommonModule, NavbarComponent ],
+  imports: [FormsModule, CommonModule, NavbarComponent],
   templateUrl: './posts.component.html',
-  styleUrls: ['./posts.component.css']
+  styleUrls: ['./posts.component.css'],
 })
 export class PostsComponent implements OnInit {
   posts$: Observable<Post[]>;
   newPostContent: string = '';
-  profilePictureUrl = 'https://cdn2.iconfinder.com/data/icons/circle-avatars-1/128/050_girl_avatar_profile_woman_suit_student_officer-512.png';
+  profilePictureUrl =
+    'https://cdn2.iconfinder.com/data/icons/circle-avatars-1/128/050_girl_avatar_profile_woman_suit_student_officer-512.png';
   loggedInUser: string | null = null;
   editingPost: Post | null = null;
 
@@ -30,7 +34,6 @@ export class PostsComponent implements OnInit {
   ngOnInit(): void {
     this.loadPosts();
     this.loggedInUser = this.authService.getUserName();
-    // console.log('Logged-in user:', this.loggedInUser);
   }
 
   loadPosts(): void {
@@ -43,8 +46,10 @@ export class PostsComponent implements OnInit {
         id: this.editingPost ? this.editingPost.id : Date.now(),
         content: this.newPostContent,
         createdBy: this.loggedInUser || 'Unknown',
-        profilePicture: this.profilePictureUrl
+        profilePicture: this.profilePictureUrl,
+        createdAt: new Date(),
       };
+      console.log('Creating/Updating post:', newPost); 
       if (this.editingPost) {
         this.store.dispatch(PostActions.updatePost({ post: newPost }));
       } else {
@@ -62,7 +67,21 @@ export class PostsComponent implements OnInit {
 
   deletePost(id: number | undefined): void {
     if (id !== undefined) {
-      this.store.dispatch(PostActions.deletePost({ id }));
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You will not be able to recover this post!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, keep it',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.store.dispatch(PostActions.deletePost({ id }));
+          Swal.fire('Deleted!', 'Your post has been deleted.', 'success');
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire('Cancelled', 'Your post is safe :)', 'error');
+        }
+      });
     } else {
       console.error('Post ID is undefined');
     }
@@ -71,5 +90,9 @@ export class PostsComponent implements OnInit {
   cancelEdit(): void {
     this.editingPost = null;
     this.newPostContent = '';
+  }
+
+  formatDate(date: Date | undefined): string {
+    return date ? format(date, 'MMMM d, yyyy h:mm a') : 'Unknown date';
   }
 }
